test(infra): promote async_test_user to shared conftest
Move the per-test async User fixture from
``tests/test_memory_db_async.py`` to ``tests/conftest.py`` so the
remaining store-conversion PRs (#1151, #1152, #1154, #1155, #1156,
#1157, #1175) can reuse it instead of redefining the same fixture
in each per-store test file. Behavior is unchanged: the fixture still
inserts a User row through the ``async_db`` connection (so the row is
visible to async store reads under READ COMMITTED) and relies on the
outer-transaction rollback for cleanup.
